// src/sdks/systemV2ApiDocs.ts
export type getSystemDictdataPageProps = {
  dictLabel?: string | null | undefined;
  dictType: string;
  dictValue?: string | null | undefined;
  pageNum: number;
  pageSize: number;
};
// ...
export interface DictTypeResp {
  dictName?: string;

  dictType?: string;

  id?: number;
}